avatar

+++ GameControl.h
@@ -11,9 +11,9 @@ struct _Point
 {
 	int X, Y, opt; // tọa độ x, y trên màn hình và opt = 1 nếu là x, opt = -1 nếu là o
 };
-long long consecutiveScore(int len, int openEnds, bool currentTurn);
-long long get_score(int a[], int player);
+int consecutiveScore(int len, int openEnds, bool currentTurn);
+int get_score(int a[], int player);
 bool valid(int x, int y); // tọa độ có hợp lệ hay không
-long long evaluation(int player); // trả về điểm trạng thái
+int evaluation(int player); // trả về điểm trạng thái
 void startGame(bool isNewGame, int XX, int YY, string name1, string name2, int avatarP1, int avatarP2, vector<pair<int, int>> Data, string fileName, int Xscore, int Oscore);
 #endif
